refactor(StopLiveTable): Extract component. Unconnect StopScheduleTable.
