+++ lib/components/viewers/stop-schedule-table.js
@@ -1,6 +1,4 @@
-import coreUtils from '@opentripplanner/core-utils'
 import React, { Component, createRef } from 'react'
-import { connect } from 'react-redux'
 import styled from 'styled-components'
 
 import { isBlank } from '../../util/ui'
@@ -13,8 +11,6 @@ import {
   stopTimeComparator
 } from '../../util/viewer'
 import StopTimeCell from './stop-time-cell'
-
-const { getTimeFormat } = coreUtils.time
 
 // Styles for the schedule table and its contents.
 const StyledTable = styled.table`
@@ -70,7 +66,7 @@ class StopScheduleTable extends Component {
   }
 
   render () {
-    const { homeTimezone, stopData, stopViewerArriving, timeFormat } = this.props
+    const { homeTimezone, stopData, timeFormat } = this.props
     const stopTimesByPattern = getStopTimesByPattern(stopData)
 
     // Merge stop times, so that we can sort them across all route patterns.
@@ -138,7 +134,6 @@ class StopScheduleTable extends Component {
                   <StopTimeCell
                     homeTimezone={homeTimezone}
                     showIcon={false}
-                    soonText={stopViewerArriving}
                     stopTime={stopTime}
                     timeFormat={timeFormat}
                     useCountdown={false}
@@ -154,19 +149,4 @@ class StopScheduleTable extends Component {
   }
 }
 
-// connect to redux store
-
-const mapStateToProps = (state, ownProps) => {
-  const { config, transitIndex, ui } = state.otp
-  return {
-    homeTimezone: config.homeTimezone,
-    stopData: transitIndex.stops[ui.viewedStop.stopId],
-    stopViewerArriving: config.language.stopViewerArriving,
-    timeFormat: getTimeFormat(config)
-  }
-}
-
-const mapDispatchToProps = {
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(StopScheduleTable)
+export default StopScheduleTable
