+++ lib/components/viewers/stop-viewer.js
@@ -9,21 +9,15 @@ import { connect } from 'react-redux'
 import styled from 'styled-components'
 
 import Icon from '../narrative/icon'
-import { setMainPanelContent, toggleAutoRefresh } from '../../actions/ui'
-import { findStop, findStopTimesForStop } from '../../actions/api'
-import { forgetStop, rememberStop, setLocation } from '../../actions/map'
-import PatternRow from './pattern-row'
+import * as uiActions from '../../actions/ui'
+import * as apiActions from '../../actions/api'
+import * as mapActions from '../../actions/map'
+import StopLiveTable from './stop-live-table'
 import StopScheduleTable from './stop-schedule-table'
 import { getShowUserSettings, getStopViewerConfig } from '../../util/state'
-import {
-  getRouteIdForPattern,
-  getStopTimesByPattern,
-  routeIsValid
-} from '../../util/viewer'
 
 const {
   getTimeFormat,
-  getUserTimezone,
   OTP_API_DATE_FORMAT
 } = coreUtils.time
 
@@ -64,47 +58,9 @@ class StopViewer extends Component {
 
   _onClickPlanFrom = () => this._setLocationFromStop('from')
 
-  _refreshStopTimes = () => {
-    const { findStopTimesForStop, viewedStop } = this.props
-    findStopTimesForStop({ stopId: viewedStop.stopId })
-    // TODO: GraphQL approach would just call findStop again.
-    // findStop({ stopId: viewedStop.stopId })
-    this.setState({ spin: true })
-    window.setTimeout(this._stopSpin, 1000)
-  }
-
-  _onToggleAutoRefresh = () => {
-    const { autoRefreshStopTimes, toggleAutoRefresh } = this.props
-    if (autoRefreshStopTimes) {
-      toggleAutoRefresh(false)
-    } else {
-      // Turn on auto-refresh and refresh immediately to give user feedback.
-      this._refreshStopTimes()
-      toggleAutoRefresh(true)
-    }
-  }
-
-  _stopSpin = () => this.setState({ spin: false })
-
   componentDidMount () {
     // Load the viewed stop in the store when the Stop Viewer first mounts
     this.props.findStop({ stopId: this.props.viewedStop.stopId })
-    // Turn on stop times refresh if enabled.
-    if (this.props.autoRefreshStopTimes) this._startAutoRefresh()
-  }
-
-  componentWillUnmount () {
-    // Turn off auto refresh unconditionally (just in case).
-    this._stopAutoRefresh()
-  }
-
-  _startAutoRefresh = () => {
-    const timer = window.setInterval(this._refreshStopTimes, 10000)
-    this.setState({ timer })
-  }
-
-  _stopAutoRefresh = () => {
-    window.clearInterval(this.state.timer)
   }
 
   _toggleFavorite = () => {
@@ -125,14 +81,8 @@ class StopViewer extends Component {
   _toggleScheduleView = () => {
     const {date, scheduleView: isShowingScheduleView} = this.state
     if (!isShowingScheduleView) {
-      // If not currently showing schedule view, fetch schedules for current
-      // date and turn off auto refresh.
-      this._stopAutoRefresh()
+      // If not currently showing schedule view, fetch schedules for current date.
       this._findStopTimesForDate(date)
-    } else {
-      // Otherwise, turn on auto refresh.
-      this._startAutoRefresh()
-      this._refreshStopTimes()
     }
     this.setState({scheduleView: !isShowingScheduleView})
   }
@@ -140,8 +90,6 @@ class StopViewer extends Component {
   _isFavorite = () => this.props.stopData &&
     this.props.favoriteStops.findIndex(s => s.id === this.props.stopData.id) !== -1
 
-  // refresh the stop in the store if the viewed stop changes w/ the
-  // Stop Viewer already mounted
   componentDidUpdate (prevProps) {
     if (
       prevProps.viewedStop &&
@@ -152,16 +100,7 @@ class StopViewer extends Component {
       this.setState(defaultState)
       this.props.findStop({ stopId: this.props.viewedStop.stopId })
     }
-    // Handle stopping or starting the auto refresh timer.
-    if (prevProps.autoRefreshStopTimes && !this.props.autoRefreshStopTimes) this._stopAutoRefresh()
-    else if (!prevProps.autoRefreshStopTimes && this.props.autoRefreshStopTimes) this._startAutoRefresh()
   }
-
-  /**
-   * Get today at midnight (morning) in seconds since epoch.
-   * FIXME: handle timezone diffs?
-   */
-  _getStartTimeForDate = date => moment(date).startOf('day').unix()
 
   handleDateChange = evt => {
     const date = evt.target.value
@@ -257,91 +196,45 @@ class StopViewer extends Component {
 
   render () {
     const {
+      autoRefreshStopTimes,
+      findStopTimesForStop,
       homeTimezone,
       stopData,
       stopViewerArriving,
       stopViewerConfig,
       timeFormat,
-      transitOperators
+      toggleAutoRefresh,
+      transitOperators,
+      viewedStop
     } = this.props
-    const { scheduleView, spin } = this.state
+    const { scheduleView } = this.state
     const hasStopTimesAndRoutes = !!(stopData && stopData.stopTimes && stopData.stopTimes.length > 0 && stopData.routes)
-
-    // construct a lookup table mapping pattern (e.g. 'ROUTE_ID-HEADSIGN') to
-    // an array of stoptimes
-    const stopTimesByPattern = getStopTimesByPattern(stopData)
-    const routeComparator = coreUtils.route.makeRouteComparator(
-      transitOperators
-    )
-    const patternHeadsignComparator = coreUtils.route.makeStringValueComparator(
-      pattern => pattern.pattern.headsign
-    )
-    const patternComparator = (patternA, patternB) => {
-      // first sort by routes
-      const routeCompareValue = routeComparator(
-        patternA.route,
-        patternB.route
-      )
-      if (routeCompareValue !== 0) return routeCompareValue
-
-      // if same route, sort by headsign
-      return patternHeadsignComparator(patternA, patternB)
-    }
 
     let contents
     if (!hasStopTimesAndRoutes) {
       contents = <div>No stop times found for date.</div>
     } else if (scheduleView) {
-      contents = <StopScheduleTable />
+      contents = (
+        <StopScheduleTable
+          homeTimezone={homeTimezone}
+          stopData={stopData}
+          timeFormat={timeFormat}
+        />
+      )
     } else {
       contents = (
-        <>
-          <div>
-            {Object.values(stopTimesByPattern)
-              .sort(patternComparator)
-              .map(({ id, pattern, route, times }) => {
-                // Only add pattern if route info is returned by OTP.
-                return routeIsValid(route, getRouteIdForPattern(pattern))
-                  ? (
-                    <PatternRow
-                      pattern={pattern}
-                      route={route}
-                      stopTimes={times}
-                      stopViewerConfig={stopViewerConfig}
-                      key={id}
-                      stopViewerArriving={stopViewerArriving}
-                      homeTimezone={homeTimezone}
-                      timeFormat={timeFormat}
-                    />
-                  )
-                  : null
-              })
-            }
-          </div>
-          {/* Auto update controls for realtime arrivals */}.
-          <div style={{ marginTop: '20px' }}>
-            <label style={{ fontWeight: 300, fontSize: 'small' }}>
-              <input
-                name='autoUpdate'
-                type='checkbox'
-                checked={this.props.autoRefreshStopTimes}
-                onChange={this._onToggleAutoRefresh} />{' '}
-                      Auto-refresh arrivals?
-            </label>
-            <button
-              className='link-button pull-right'
-              style={{ fontSize: 'small' }}
-              onClick={this._refreshStopTimes}>
-              <Icon
-                className={spin ? 'fa-spin' : ''}
-                type='refresh' />{' '}
-              {moment(stopData.stopTimesLastUpdated)
-                .tz(getUserTimezone())
-                .format(timeFormat)}
-            </button>
-          </div>
-          }
-        </>
+        <StopLiveTable
+          autoRefreshStopTimes={autoRefreshStopTimes}
+          findStopTimesForStop={findStopTimesForStop}
+          homeTimezone={homeTimezone}
+          stopData={stopData}
+          stopViewerArriving={stopViewerArriving}
+          stopViewerConfig={stopViewerConfig}
+          timeFormat={timeFormat}
+          transitOperators={transitOperators}
+          toggleAutoRefresh={toggleAutoRefresh}
+          viewedStop={viewedStop}
+        />
       )
     }
 
@@ -384,13 +277,13 @@ const mapStateToProps = (state, ownProps) => {
 }
 
 const mapDispatchToProps = {
-  findStop,
-  findStopTimesForStop,
-  forgetStop,
-  rememberStop,
-  setLocation,
-  setMainPanelContent,
-  toggleAutoRefresh
+  findStop: apiActions.findStop,
+  findStopTimesForStop: apiActions.findStopTimesForStop,
+  forgetStop: mapActions.forgetStop,
+  rememberStop: mapActions.rememberStop,
+  setLocation: mapActions.setLocation,
+  setMainPanelContent: uiActions.setMainPanelContent,
+  toggleAutoRefresh: uiActions.toggleAutoRefresh
 }
 
 export default connect(mapStateToProps, mapDispatchToProps)(StopViewer)
